requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
pycocotools>=2.0.6
pycocoevalcap>=1.2
datasets>=2.14.0
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0
nvidia-ml-py3>=7.352.0
